DO NOT MERGE: Revert "Move initial coverage clearing out of the test type."
Revert submission 12789043-host-test-coverage

Reason for revert: Coverage tests not generating artifacts
Reverted Changes:
I61dde583c:Move initial coverage clearing out of the test typ...
If7479c00a:Add ClangCodeCoverageCollector as an AutoLogCollec...
I1db451d43:Change ClangCodeCoverageListener to implement Base...

+++ src/com/android/tradefed/device/metric/JavaCodeCoverageCollector.java
@@ -24,16 +24,13 @@ import com.android.tradefed.config.IConfigurationReceiver;
 import com.android.tradefed.config.Option;
 import com.android.tradefed.device.DeviceNotAvailableException;
 import com.android.tradefed.device.ITestDevice;
-import com.android.tradefed.invoker.IInvocationContext;
 import com.android.tradefed.invoker.logger.CurrentInvocation;
 import com.android.tradefed.metrics.proto.MetricMeasurement.Metric;
 import com.android.tradefed.result.FailureDescription;
 import com.android.tradefed.result.FileInputStreamSource;
 import com.android.tradefed.result.LogDataType;
-import com.android.tradefed.result.ITestInvocationListener;
 import com.android.tradefed.result.error.InfraErrorIdentifier;
 import com.android.tradefed.testtype.coverage.CoverageOptions;
-import com.android.tradefed.util.AdbRootElevator;
 import com.android.tradefed.util.FileUtil;
 import com.android.tradefed.util.JavaCodeCoverageFlusher;
 
@@ -73,22 +70,6 @@ public final class JavaCodeCoverageCollector extends BaseDeviceMetricCollector
     private IConfiguration mConfiguration;
 
     @Override
-    public ITestInvocationListener init(
-            IInvocationContext context, ITestInvocationListener listener) {
-        super.init(context, listener);
-
-        if (isJavaCoverageEnabled()) {
-            try (AdbRootElevator adbRoot = new AdbRootElevator(getDevices().get(0))) {
-                getCoverageFlusher().resetCoverage();
-            } catch (DeviceNotAvailableException e) {
-                throw new RuntimeException(e);
-            }
-        }
-
-        return this;
-    }
-
-    @Override
     public void setConfiguration(IConfiguration configuration) {
         mConfiguration = configuration;
     }
@@ -115,7 +96,12 @@ public final class JavaCodeCoverageCollector extends BaseDeviceMetricCollector
 
     @Override
     public void onTestRunEnd(DeviceMetricData runData, final Map<String, Metric> runMetrics) {
-        if (!isJavaCoverageEnabled()) {
+        if (mConfiguration == null
+                || !mConfiguration.getCoverageOptions().isCoverageEnabled()
+                || !mConfiguration
+                        .getCoverageOptions()
+                        .getCoverageToolchains()
+                        .contains(CoverageOptions.Toolchain.JACOCO)) {
             return;
         }
         if (MERGE_COVERAGE_MEASUREMENTS_TEST_NAME.equals(getRunName())) {
@@ -152,7 +138,7 @@ public final class JavaCodeCoverageCollector extends BaseDeviceMetricCollector
             ImmutableList.Builder<String> devicePaths = ImmutableList.builder();
             devicePaths.add(testCoveragePath);
 
-            try (AdbRootElevator adbRoot = new AdbRootElevator(device)) {
+            try {
                 if (mConfiguration.getCoverageOptions().isCoverageFlushEnabled()) {
                     getCoverageFlusher().forceCoverageFlush();
                 }
@@ -161,7 +147,7 @@ public final class JavaCodeCoverageCollector extends BaseDeviceMetricCollector
                 String fileList = device.executeShellCommand(FIND_COVERAGE_FILES);
                 devicePaths.addAll(Splitter.on('\n').omitEmptyStrings().split(fileList));
 
-                collectAndLogCoverageMeasurements(device, devicePaths.build());
+                collectAndLogCoverageMeasurementsAsRoot(device, devicePaths.build());
 
             } catch (DeviceNotAvailableException | IOException e) {
                 throw new RuntimeException(e);
@@ -175,12 +161,39 @@ public final class JavaCodeCoverageCollector extends BaseDeviceMetricCollector
         }
     }
 
+    private void collectAndLogCoverageMeasurementsAsRoot(
+            ITestDevice device, List<String> devicePaths)
+            throws IOException, DeviceNotAvailableException {
+
+        // We enable root before pulling files off the device since the coverage file of the test
+        // process is written in its private directory which is otherwise inaccessible. Coverage
+        // files of other processes should be accessible without root. Note that we also restore
+        // root status to what it was after we're done to not interfere with subsequent tests that
+        // run on the device.
+        boolean wasRoot = device.isAdbRoot();
+        if (!wasRoot && !device.enableAdbRoot()) {
+            throw new RuntimeException(
+                    "Failed to enable root before pulling Java code coverage files off device");
+        }
+
+        try {
+            collectAndLogCoverageMeasurements(device, devicePaths);
+        } finally {
+            for (String devicePath : devicePaths) {
+                device.deleteFile(devicePath);
+            }
+            if (!wasRoot && !device.disableAdbRoot()) {
+                throw new RuntimeException(
+                        "Failed to disable root after pulling Java code coverage files off device");
+            }
+        }
+    }
+
     private void collectAndLogCoverageMeasurements(ITestDevice device, List<String> devicePaths)
             throws IOException, DeviceNotAvailableException {
 
         for (String devicePath : devicePaths) {
             File coverageFile = device.pullFile(devicePath);
-            device.deleteFile(devicePath);
             verifyNotNull(
                     coverageFile, "Failed to pull the Java code coverage file from %s", devicePath);
 
@@ -205,14 +218,5 @@ public final class JavaCodeCoverageCollector extends BaseDeviceMetricCollector
 
     private FailureDescription createCodeCoverageFailure(String message) {
         return CurrentInvocation.createFailure(message, InfraErrorIdentifier.CODE_COVERAGE_ERROR);
-    }
-
-    private boolean isJavaCoverageEnabled() {
-        return mConfiguration != null
-                && mConfiguration.getCoverageOptions().isCoverageEnabled()
-                && mConfiguration
-                        .getCoverageOptions()
-                        .getCoverageToolchains()
-                        .contains(CoverageOptions.Toolchain.JACOCO);
     }
 }
